REFACTOR - Profile Image Component & minor fix
